requirements.txt
flask
keras
tensorflow
pillow